+++ src/main/java/ru/pnzgu/springblog/services/impl/PostServiceImpl.java
@@ -3,11 +3,13 @@ package ru.pnzgu.springblog.services.impl;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 import ru.pnzgu.springblog.dto.common.PageDto;
 import ru.pnzgu.springblog.dto.post.CreatePostRequest;
 import ru.pnzgu.springblog.dto.post.GetPostResponse;
 import ru.pnzgu.springblog.dto.post.UpdatePostRequest;
+import ru.pnzgu.springblog.exceptions.AccessDeniedException;
 import ru.pnzgu.springblog.exceptions.EntityNotFoundException;
 import ru.pnzgu.springblog.exceptions.ValidationException;
 import ru.pnzgu.springblog.helpers.AuthFacade;
@@ -110,6 +112,12 @@ public class PostServiceImpl implements PostService {
     @Override
     @Transactional
     public GetPostResponse publish(int id) {
+        var authorities = authFacade.getAuth().getAuthorities();
+        
+        if (!authorities.contains(new SimpleGrantedAuthority("ADMIN")) && 
+                !authorities.contains(new SimpleGrantedAuthority("MODERATOR")))
+            throw new AccessDeniedException("Access denied");
+        
         var post = postRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Post not found"));
         
         if (post.isPublished())
@@ -125,7 +133,16 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public void delete(int id) {
+        var authorities = authFacade.getAuth().getAuthorities();
+        var username = authFacade.getAuth().getName();
+        var currentUser = userRepository.findByUsername(username).orElseThrow(() -> new EntityNotFoundException("User not found"));
         var post = postRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Post not found"));
+
+        if (!authorities.contains(new SimpleGrantedAuthority("ADMIN")) &&
+                !authorities.contains(new SimpleGrantedAuthority("MODERATOR")) &&
+                currentUser.getId() != post.getUser().getId())
+            throw new AccessDeniedException("Access denied");
+        
         postRepository.delete(post);
     }
